Move `primitive_structure_from_cif` to own module
In `aiida-core==1.0.0`, the source code of the file in which the
function is defined is copied entirely to the repository. To prevent an
unnecessary bloating of the repository one should ensure that only
relevant code is contained in the file of the function. To this end it
is best to keep each process function within its own file.
